Implment HAL interface for vendor specific commands/events
Associated change in hardware/libhardware introduces new
API for sending vendor specific command to the controller
and receiving vendor specific commands from the controller.

This chage adds code to btif to hook up the API to the existing
functionality that exists within the BTM stack layer.

This change is pending approval on the AOSP gerrit.

Change-Id: I12f27c8b13fae8685412851f2b0cac2152bec330
Signed-off-by: James Bootsma <james.bootsma@garmin.com>